+++ demisto_sdk/commands/content_graph/interface/graph.py
@@ -114,7 +114,10 @@ class ContentGraphInterface(ABC):
 
     @abstractmethod
     def find_core_packs_depend_on_non_core_packs(
-        self, pack_ids: List[str], core_pack_list
+        self,
+        pack_ids: List[str],
+        marketplace: MarketplaceVersions,
+        core_pack_list: List[str],
     ) -> List[BaseContent]:
         pass