+++ demisto_sdk/tests/integration_tests/upload_integration_test.py
@@ -1,3 +1,5 @@
+import shutil
+import tempfile
 from os.path import join
 
 import click
@@ -6,10 +8,17 @@ from click.testing import CliRunner
 from packaging.version import parse
 
 from demisto_sdk.__main__ import main
+from demisto_sdk.commands.common.constants import GENERAL_DEFAULT_FROMVERSION
+from demisto_sdk.commands.common.handlers import JSON_Handler, YAML_Handler
 from demisto_sdk.commands.common.legacy_git_tools import git_path
+from TestSuite.test_tools import ChangeCWD
 
 UPLOAD_CMD = "upload"
 DEMISTO_SDK_PATH = join(git_path(), "demisto_sdk")
+
+
+json = JSON_Handler()
+yaml = YAML_Handler()
 
 
 @pytest.fixture
@@ -32,7 +41,7 @@ def demisto_client(mocker):
     mocker.patch("click.secho")
 
 
-def test_integration_upload_pack_positive(demisto_client, mocker, repo):
+def test_integration_upload_pack_positive(demisto_client, repo):
     """
     Given
     - Content pack named FeedAzure to upload.
@@ -44,14 +53,6 @@ def test_integration_upload_pack_positive(demisto_client, mocker, repo):
     - Ensure upload runs successfully.
     - Ensure success upload message is printed.
     """
-    from demisto_sdk.commands.content_graph.objects.integration_script import (
-        IntegrationScript,
-    )
-
-    mocker.patch.object(
-        IntegrationScript, "get_supported_native_images", return_value=[]
-    )
-
     pack_path = join(
         DEMISTO_SDK_PATH, "tests/test_files/content_repo_example/Packs/FeedAzure"
     )
@@ -89,6 +90,66 @@ def test_integration_upload_pack_positive(demisto_client, mocker, repo):
     )
 
     assert not result.stderr
+
+
+def test_zipped_pack_upload_positive(repo, mocker, demisto_client):
+    """
+    Given
+    - content pack name
+
+    When
+    - Uploading the zipped pack.
+
+    Then
+    - Ensure upload runs successfully.
+    - Ensure success upload message is printed.
+    - ensure yml / json content items inside the pack are getting unified.
+    """
+    pack = repo.setup_one_pack(name="test-pack")
+
+    mocker.patch(
+        "demisto_sdk.commands.upload.uploader.Uploader.notify_user_should_override_packs",
+        return_value=True,
+    )
+
+    mocker.patch(
+        "demisto_sdk.commands.common.content.objects.pack_objects.pack.get_demisto_version",
+        return_value=parse(GENERAL_DEFAULT_FROMVERSION),
+    )
+
+    runner = CliRunner(mix_stderr=False)
+    with tempfile.TemporaryDirectory() as dir:
+        with ChangeCWD(pack.repo_path):
+            result = runner.invoke(
+                main,
+                [UPLOAD_CMD, "-i", pack.path, "-z", "--insecure", "--keep-zip", dir],
+            )
+
+        shutil.unpack_archive(f"{dir}/uploadable_packs.zip", dir, "zip")
+        shutil.unpack_archive(f"{dir}/test-pack.zip", dir, "zip")
+
+        with open(
+            f"{dir}/Layouts/layoutscontainer-test-pack_layoutcontainer.json"
+        ) as file:
+            layout_content = json.load(file)
+
+        # validate json based content entities are being unified before getting zipped
+        assert "fromServerVersion" in layout_content
+        assert "toServerVersion" in layout_content
+
+        with open(f"{dir}/Integrations/integration-test-pack_integration.yml") as file:
+            integration_content = yaml.load(file)
+
+        # validate yml based content entities are being unified before getting zipped
+        assert "nativeimage" in integration_content.get("script", {})
+
+    assert result.exit_code == 0
+
+    assert "SUCCESSFUL UPLOADS:" in click.secho.call_args_list[4][0][0]
+    assert (
+        "╒═══════════╤════════╕\n│ NAME      │ TYPE   │\n╞═══════════╪════════╡\n│ test-pack │ pack   │\n╘═══════════╧════════╛\n"
+        in click.secho.call_args_list[5][0][0]
+    )
 
 
 def test_integration_upload_path_does_not_exist(demisto_client):